better docstrings

+++ src/pars.jl
@@ -42,8 +42,7 @@ as the zero vector. Now every `adaptskip` iterations, `mT` is updated to the val
     κ :: Float64 = 100.0 # in variance of prior on initial momenta
     δpos :: Float64 = 0.001
     δmom :: Float64 = 0.1
-    #δinit ::Array{Float64,1} = [0.001, 0.1]
-    skip_saveITER :: Float64 = 10 # skip every skip_saveITER iteration in saving paths
+    skip_saveITER :: Int64 = 10 # skip every skip_saveITER iteration in saving paths
 end
 
 @with_kw struct Pars_ahs <: Pars
@@ -62,6 +61,5 @@ end
     κ :: Float64 = 100.0 # in variance of prior on initial momenta    δinit_ms ::Array{Float64,1} = [0.001, 0.1]
     δpos :: Float64 = 0.1
     δmom :: Float64 = 0.1
-    #δinit ::Array{Float64,1} = [0.1, 0.1]
-    skip_saveITER :: Float64 = 10 # skip every skip_saveITER iteration in saving paths
+    skip_saveITER :: Int64 = 10 # skip every skip_saveITER iteration in saving paths
 end
